requirements.txt
amqp==5.2.0
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
async-timeout==4.0.3
billiard==4.2.0
cachetools==5.3.3
certifi==2024.2.2
chardet==3.0.4
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cron-descriptor==1.4.3
distro==1.9.0
Django==4.2
django-admin-rangefilter==0.13.0
django-admin-volt==1.0.10
django-extensions==3.2.3
django-modeltranslation==0.19.2
django-redis==5.4.0
django-timezone-field==6.1.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
exceptiongroup==1.2.1
Faker==25.0.1
google-api-core==2.19.0
google-auth==2.29.0
google-cloud-core==2.4.1
google-cloud-translate==3.15.3
googleapis-common-protos==1.63.0
grpcio==1.64.0
grpcio-status==1.62.2
h11==0.14.0
h2==3.2.0
hpack==3.0.0
hstspreload==2024.5.1
httpcore==1.0.5
httpx==0.27.0
hyperframe==5.2.0
idna==3.7
kombu==5.3.7
numpy==1.26.4
openai==1.30.3
opencv-python==4.9.0.80
pillow==10.3.0
prompt_toolkit==3.0.45
proto-plus==1.23.0
protobuf==4.25.3
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.7.1
pydantic_core==2.18.2
PyJWT==2.8.0
python-crontab==3.1.0
python-dateutil==2.9.0.post0
python-decouple==3.8
redis==5.0.4
requests==2.32.2
rfc3986==1.5.0
rsa==4.9
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.0
toposort==1.10
tqdm==4.66.4
typing_extensions==4.12.0
tzdata==2024.1
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.13
